--- a/src/Controls/src/Core/BindableObject.cs
+++ b/src/Controls/src/Core/BindableObject.cs
@@ -124,8 +124,8 @@ namespace Microsoft.Maui.Controls
 			if (bpcontext == null)
 				return;
 
-			var original = bpcontext.Values.LastOrDefault().Value;
-			var newValue = bpcontext.Values.Count >= 2 ? bpcontext.Values[bpcontext.Values.Keys[bpcontext.Values.Count - 2]] : null;
+			var original = bpcontext.Values.GetSpecificityAndValue().Value;
+			var newValue = bpcontext.Values.GetClearedValue();
 			var changed = !Equals(original, newValue);
 			if (changed)
 			{
@@ -158,7 +158,7 @@ namespace Microsoft.Maui.Controls
 
 			var context = property.DefaultValueCreator != null ? GetOrCreateContext(property) : GetContext(property);
 
-			return context == null ? property.DefaultValue : context.Values.Last().Value;
+			return context == null ? property.DefaultValue : context.Values.GetSpecificityAndValue().Value;
 		}
 
 		internal LocalValueEnumerator GetLocalValueEnumerator() => new LocalValueEnumerator(this);
@@ -175,7 +175,7 @@ namespace Microsoft.Maui.Controls
 			{
 				if (_propertiesEnumerator.MoveNext())
 				{
-					Current = new LocalValueEntry(_propertiesEnumerator.Current.Key, _propertiesEnumerator.Current.Value.Values.LastOrDefault().Value, _propertiesEnumerator.Current.Value.Attributes);
+					Current = new LocalValueEntry(_propertiesEnumerator.Current.Key, _propertiesEnumerator.Current.Value.Values.GetSpecificityAndValue().Value, _propertiesEnumerator.Current.Value.Attributes);
 					return true;
 				}
 				return false;
@@ -213,8 +213,9 @@ namespace Microsoft.Maui.Controls
 			{
 				if (properties.TryGetValue(propArray[i], out var context))
 				{
-					resultArray[i].IsSet = context.Values.LastOrDefault().Key.CompareTo(SetterSpecificity.DefaultValue) != 0;
-					resultArray[i].Value = (T)context.Values.LastOrDefault().Value;
+					var pair = context.Values.GetSpecificityAndValue();
+					resultArray[i].IsSet = pair.Key.CompareTo(SetterSpecificity.DefaultValue) != 0;
+					resultArray[i].Value = (T)pair.Value;
 				}
 				else
 				{
@@ -239,7 +240,7 @@ namespace Microsoft.Maui.Controls
 				return false;
 			if ((bpcontext.Attributes & BindableContextAttributes.IsDefaultValueCreated) == BindableContextAttributes.IsDefaultValueCreated)
 				return true;
-			return bpcontext.Values.LastOrDefault().Key.CompareTo(SetterSpecificity.DefaultValue) != 0;
+			return bpcontext.Values.GetSpecificityAndValue().Key.CompareTo(SetterSpecificity.DefaultValue) != 0;
 		}
 
 
@@ -290,9 +291,9 @@ namespace Microsoft.Maui.Controls
 			var context = GetOrCreateContext(targetProperty);
 
 			//if the value is manually set (has highest specificity than FromBinding), we reassign the specificity so it'll get replaced when the binding is applied
-			if (context.Values.Last().Key.CompareTo(SetterSpecificity.FromBinding) > 0)
+			var kvp = context.Values.GetSpecificityAndValue();
+			if (kvp.Key.CompareTo(SetterSpecificity.FromBinding) > 0)
 			{
-				var kvp = context.Values.Last();
 				context.Values.Remove(kvp.Key);
 				context.Values[SetterSpecificity.FromBinding] = kvp.Value;
 			}
@@ -332,16 +333,11 @@ namespace Microsoft.Maui.Controls
 		{
 			//I wonder if we coulnd't treat bindingcoutext with specificities
 			BindablePropertyContext bpContext = bindable.GetContext(BindingContextProperty);
-			if (bpContext != null && bpContext.Values.LastOrDefault().Key.CompareTo(SetterSpecificity.ManualValueSetter) >= 0)
+			if (bpContext != null && bpContext.Values.GetSpecificityAndValue().Key.CompareTo(SetterSpecificity.ManualValueSetter) >= 0)
 				return;
 
-			object oldContext = bindable._inheritedContext?.Target;
-
-			if (ReferenceEquals(oldContext, value))
+			if (ReferenceEquals(bindable._inheritedContext?.Target, value))
 				return;
-
-			if (bpContext != null && oldContext == null)
-				oldContext = bpContext.Values.LastOrDefault().Value;
 
 			var binding = bpContext?.Bindings.Values.LastOrDefault();
 
@@ -580,15 +576,17 @@ namespace Microsoft.Maui.Controls
 
 		void SetValueActual(BindableProperty property, BindablePropertyContext context, object value, bool currentlyApplying, SetValueFlags attributes, SetterSpecificity specificity, bool silent = false)
 		{
-			object original = context.Values.LastOrDefault().Value;
-			var originalSpecificity = context.Values.LastOrDefault().Key;
+			var pair = context.Values.GetSpecificityAndValue();
+			var original = pair.Value;
+			var originalSpecificity = pair.Key;
 
 			//if the last value was set from handler, override it
 			if (specificity != SetterSpecificity.FromHandler
 				&& originalSpecificity == SetterSpecificity.FromHandler)
 			{
 				context.Values.Remove(SetterSpecificity.FromHandler);
-				originalSpecificity = context.Values.LastOrDefault().Key;
+				pair = context.Values.GetSpecificityAndValue();
+				originalSpecificity = pair.Key;
 			}
 
 			//We keep setter of lower specificity so we can unapply
@@ -681,7 +679,7 @@ namespace Microsoft.Maui.Controls
 		{
 			var defaultValueCreator = property.DefaultValueCreator;
 			var context = new BindablePropertyContext { Property = property };
-			context.Values.Add(SetterSpecificity.DefaultValue, defaultValueCreator != null ? defaultValueCreator(this) : property.DefaultValue);
+			context.Values.SetValue(SetterSpecificity.DefaultValue, defaultValueCreator != null ? defaultValueCreator(this) : property.DefaultValue);
 
 			if (defaultValueCreator != null)
 				context.Attributes = BindableContextAttributes.IsDefaultValueCreated;
@@ -763,7 +761,7 @@ namespace Microsoft.Maui.Controls
 			if (bpcontext == null)
 				return;
 
-			object currentValue = bpcontext.Values.LastOrDefault().Value;
+			object currentValue = bpcontext.Values.GetSpecificityAndValue().Value;
 
 			if (property.ValidateValue != null && !property.ValidateValue(this, currentValue))
 				throw new ArgumentException($"Value is an invalid value for {property.PropertyName}", nameof(currentValue));
@@ -789,7 +787,7 @@ namespace Microsoft.Maui.Controls
 
 			public Queue<SetValueArgs> DelayedSetters;
 			public BindableProperty Property;
-			public SortedList<SetterSpecificity, object> Values = new();
+			public readonly SetterSpecificityList Values = new();
 		}