@jonathanpeppers 2023-09-20
This also seems to help startup, a `dotnet new maui` app on a Pixel 5 testing this change with `dotnet-trace` attached:
```
Before:
105.36ms microsoft.maui.controls!Microsoft.Maui.Controls.BindableObject.SetValueCore(Microsoft.Maui.Controls.BindableProperty...
After:
89.79ms microsoft.maui.controls!Microsoft.Maui.Controls.BindableObject.SetValueCore(Microsoft.Maui.Controls.BindableProperty..
```
So *maybe* ~15ms improved for this call?