@jonathanpeppers 2023-09-20
A customer just shared some results comparing their app startup 7 vs 8:

https://discord.com/channels/732297728826277939/732297837953679412/1154161770169114654

![image](https://github.com/dotnet/maui/assets/840039/bd28217e-ade7-48e6-ae89-0a411d6a69d2)