@StephaneDelcroix 2023-09-21
can't we get the same perf bump by just avoiding calling Last() ?