@StephaneDelcroix 2023-09-21
I'd go with that, and if you have a chart indicating where the time is spend, we can look at how to improve this further