@jonathanpeppers 2023-09-21
> can't we get the same perf bump by just avoiding calling Last() ?

Something like `list[list.Count-1]` is already done by `Last()`:

https://source.dot.net/#System.Linq/System/Linq/Last.cs,76

I think that `list[list.Count-1]` would be roughly the same as calling `Last()`. But we can try it.

Some other ideas, we could look at next:

* Invert the sorting. Is `First()` faster than `Last()`?
* Is there some other collection type that would work better?