@mattleibow 2023-09-21
Can you also just store last? So first, second and last.