@jonathanpeppers 2023-09-22
> Can you also just store last? So first, second and last.

I have tested some benchmarks, but haven't proven it 100% helps yet. I think we should do this in a future PR.